[#170] refactor: 일정 조회 로직을 query 유즈케이스가 아닌 command 유즈케이스로 간주
