+++ src/main/java/com/kiero/schedule/application/port/in/ScheduleQueryUseCase.java
@@ -4,11 +4,8 @@ import java.time.LocalDate;
 
 import com.kiero.schedule.application.dto.DefaultScheduleContentResponse;
 import com.kiero.schedule.application.dto.ScheduleTabResponse;
-import com.kiero.schedule.application.dto.TodayScheduleResponse;
 
 public interface ScheduleQueryUseCase {
-
-	TodayScheduleResponse getTodaySchedule(Long childId);
 
 	ScheduleTabResponse getSchedules(LocalDate startDate, LocalDate endDate, Long parentId, Long childId);
